[Silver III] Title: 주유소, Time: 24 ms, Memory: 3684 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -23,7 +23,7 @@ int main() {
 	int minP = price[0];
 	for (int i = 0; i < N; i++) {
 		if (minP > price[i]) minP = price[i];
-		total += minP * road[i];
+		total += (long long)minP * road[i];
 	}
 
 	cout << total;
